Added methods for deleting and updating books in interfaces and implemented in repositories and used in kotnolers, and these two methods were also covered with tests

+++ BooksWebAPI/Repositories/BookRepository.cs
@@ -1,6 +1,7 @@
 ﻿using BooksWebAPI.Data;
 using BooksWebAPI.Inerfaces;
 using BooksWebAPI.Models;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System.Security.Claims;
@@ -27,10 +28,32 @@ namespace BooksWebAPI.Repositories
             return await _context.Books.Where(book => Convert.ToString(book.UserId) == userId).ToListAsync();
         }
 
-        public bool Remove(Book book)
+        public async Task<bool> Remove(int id)
         {
-            _context.Books.Remove(book);
-            return Save();
+            var book =  await _context.Books.FirstOrDefaultAsync(book => book.Id == id);
+            if (book != null)
+            {
+                _context.Books.Remove(book);
+                return Save();
+            }
+            return false;
+        }
+
+        public async Task<bool> Update(int id, Book book)
+        {
+            if (id != book.Id)
+                return false;
+
+            try
+            {
+                _context.Entry(book).State = EntityState.Modified;
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateConcurrencyException)
+            {
+                return false;
+            }
+            return true;
         }
 
         public bool IsNull(string userId)
